Lots of improvements. See details.
I've changed to a new level of Android, since many functions were not working properly in RecyclerViews.
Dialog for asking reasons to cancel an order (not finished)
Dialog to confirm a restaurant order (not finised)
More Chinese translations
Much better layouts in xml, faster and clean organized.
More usage of GRPC messages.
I prefer C++ brackets, all below each other (matter of taste).
I still like m_ in names, but willing to change if more people want to participate in the project.
Ordering system starts to work, +1, -1, remove items,
Next phase is billing and askTable. (all is takeaway for now)
Deepseek creates lots of useful code and improvements.
Cursor class, so not all the software uses Ints.
Starting the Payment class.
Sorting order like the C++ pos computer from www.kassaku.nl
TimeFrame class, to see each partial order for a table.
Conversion of payment details.
TimeStamp class for adding times and testing timeframe.
Less global.
Callbacks between data updates and UI updates.
CookingState
Payment Method copied.
Many new layouts
Proto files slightly updated, these are converted to Java/Kotlin automatic.

+++ microfood/app/src/main/java/com/hha/service/DailyTransactionItemService.kt
@@ -320,12 +320,12 @@ class DailyTransactionItemService(channel: ManagedChannel) : BaseGrpcService<Dai
         }
     }
 
-    fun undoTimeFrame(transactionId: Long, timeFrameIndex: Int, deviceId: Int): Boolean = runBlocking {
+    fun undoTimeFrame(transactionId: Long, timeFrameIndex: Int, deviceId: Short): Boolean = runBlocking {
         try {
             val request = UndoTimeFrameRequest.newBuilder()
                 .setTransactionId(transactionId)
                 .setTimeFrameIndex(timeFrameIndex)
-                .setDeviceId(deviceId)
+                .setDeviceId(deviceId.toInt())
                 .build()
             stub.undoTimeFrame(request)
             true
